# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthDto"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/students": {
      "get": {
        "operationId": "WorkflowController_listStudents",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mine",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Workflow"
        ]
      },
      "post": {
        "operationId": "WorkflowController_createStudent",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "description": "UUID retained when retrying identical input.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Workflow"
        ]
      }
    },
...
    "/api/sessions": {
      "get": {
        "operationId": "WorkflowController_sessions",
        "parameters": [
          {
            "name": "week",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classGroupId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Workflow"
        ]
      },
      "post": {
        "operationId": "WorkflowController_createSession",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "description": "UUID retained when retrying identical input.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Workflow"
        ]
      }
    },
...
    "/api/sessions/{id}/participants": {
      "get": {
        "operationId": "WorkflowController_roster",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Workflow"
        ]
      },
      "post": {
        "operationId": "WorkflowController_add",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "description": "UUID retained when retrying identical input.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddParticipantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Workflow"
        ]
      }
    },
...
  },
  "info": {
    "title": "StudentSys API",
    "description": "Timetable, trial booking, teacher feedback and private administrator follow-up.",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "student_session"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "alice@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "TEACHER"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role"
        ]
      },
      "AuthDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "csrfToken": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "csrfToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "StudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "yearLevel": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "yearLevel"
        ]
      },
      "StudentPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "guardianName": {
            "type": "string"
          },
          "guardianRelationship": {
            "type": "string"
          },
          "guardianPhone": {
            "type": "string"
          },
          "guardianEmail": {
            "type": "string"
          },
          "guardianWechat": {
            "type": "string"
          },
          "preferredChannel": {
            "type": "string",
            "enum": [
              "",
              "EMAIL",
              "PHONE",
              "SMS",
              "WECHAT"
            ]
          },
          "preferredLanguage": {
            "type": "string",
            "enum": [
              "en-AU",
              "zh-CN"
            ]
          },
          "learningGoals": {
            "type": "string"
          },
          "preferredTimes": {
            "type": "string"
          },
          "interestedSubjects": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "yearLevel": {
            "type": "string",
            "enum": [
              "Foundation",
              "Year 1",
              "Year 2",
              "Year 3",
              "Year 4",
              "Year 5",
              "Year 6",
              "Year 7",
              "Year 8",
              "Year 9",
              "Year 10",
              "Year 11",
              "Year 12",
              "Not assessed"
            ]
          },
          "giftTrialCredit": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "yearLevel"
        ]
      },
      "LessonDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "className": {
            "type": "string"
          },
          "classGroupId": {
            "type": "string"
          },
          "targetLevel": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          },
          "courseName": {
            "type": "string"
          },
          "teacherName": {
            "type": "string"
          },
          "teacherId": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "participantCount": {
            "type": "number"
          },
          "trialCount": {
            "type": "number"
          },
          "newCount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "feedbackSubmittedAt": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "className",
          "classGroupId",
          "targetLevel",
          "courseId",
          "courseName",
          "teacherName",
          "teacherId",
          "startsAt",
          "endsAt",
          "capacity",
          "participantCount",
          "trialCount",
          "newCount",
          "status",
          "version",
          "feedbackSubmittedAt",
          "summary"
        ]
      },
...
      "ActionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "CreateSessionDto": {
        "type": "object",
        "properties": {
          "classGroupId": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          },
          "teacherId": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          }
        },
        "required": [
          "classGroupId",
          "courseId",
          "teacherId",
          "startsAt",
          "endsAt",
          "capacity"
        ]
      },
...
      "ParticipantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "yearLevel": {
            "type": "string"
          },
          "participantId": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "bookingStatus": {
            "type": "string"
          },
          "attendance": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "canManage": {
            "type": "boolean"
          },
          "feedback": {
            "type": "string",
            "nullable": true
          },
          "abilityNote": {
            "type": "string",
            "nullable": true
          },
          "preferenceNote": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "yearLevel",
          "participantId",
          "kind",
          "category",
          "bookingStatus",
          "attendance",
          "version",
          "canManage",
          "feedback",
          "abilityNote",
          "preferenceNote"
        ]
      },
      "RosterDto": {
        "type": "object",
        "properties": {
          "lesson": {
            "$ref": "#/components/schemas/LessonDto"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantDto"
            }
          },
          "cancelled": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantDto"
            }
          }
        },
        "required": [
          "lesson",
          "participants",
          "cancelled"
        ]
      },
      "AddParticipantDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "TRIAL",
              "REGULAR"
            ]
          }
        },
        "required": [
          "studentId",
          "kind"
        ]
      },
...
    }
  }
}